move actions footer to common folder
