+++ client/packages/common/src/ui/components/footer/index.ts
@@ -1,0 +1,1 @@
+export * from './ActionsFooter';
