+++ client/packages/invoices/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["./src"]
+  "include": ["./src", "../common/src/ui/components/footer/ActionsFooter"]
 }
